+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Breaking change: Migrate `HeaderNav` and remove `k-HeaderNav` styles.
 - Breaking change: Remove `FormRow` component and `k-FormRow` styles.
 - Breaking change: Remove `Search` component and `k-SearchInput` styles.
+- Breaking change: Convert `LinkList` component and delete `k-LinkList` styles.
 
 ## [3.0.0-beta.2] - 2021-02-15